# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: SprintPulse API
  version: 0.2.0
  license:
    name: Private hackathon prototype
  description: |
    OpenAPI contract for the Express routes mounted at `/api` in SprintPulse.

    The API currently uses a `personaId` query or body field to resolve the acting user. Some routes are available only
    when `ENABLE_MOCK_FLOW=true`; in Supabase data mode those routes return `501` until their adapters are implemented.
servers:
  - url: http://localhost:4000/api
    description: Local SprintPulse API
...
security: []
paths:
...
  /projects/{projectId}/sprints:
    get:
      tags: [Projects]
      operationId: listProjectSprints
      summary: List project sprints and the current sprint summary.
      parameters:
        - $ref: "#/components/parameters/ProjectIdPath"
        - $ref: "#/components/parameters/PersonaIdQuery"
      responses:
        "200":
          description: Sprint summaries for the project.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SprintListResponse"
        "404":
          description: Sprints are not found or not visible.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                notVisible:
                  value:
                    error: Sprints not found or not visible to this user
        "500":
          $ref: "#/components/responses/InternalServerError"
...
  /projects/{projectId}/integrations:
    get:
      tags: [Integrations]
      operationId: getProjectIntegrations
      summary: Get Jira, GitHub, recent sync runs, and signal previews.
      parameters:
        - $ref: "#/components/parameters/ProjectIdPath"
        - $ref: "#/components/parameters/PersonaIdQuery"
      responses:
        "200":
          description: Integration status and previews.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IntegrationStatusResponse"
        "404":
          description: Integrations are not found or not visible.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                notVisible:
                  value:
                    error: Integrations not found or not visible to this user
        "500":
          $ref: "#/components/responses/InternalServerError"
  /projects/{projectId}/jira/configure:
    post:
      tags: [Integrations]
      operationId: configureProjectJira
      summary: Configure Jira for a project.
      description: Saves Jira site and project key metadata. In mock flow this route returns `501`.
      parameters:
        - $ref: "#/components/parameters/ProjectIdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConfigureJiraRequest"
            examples:
              configureJira:
                value:
                  personaId: priya
                  jiraSite: semicolons.atlassian.net
                  projectKey: SP
      responses:
        "200":
          description: Jira connection configured.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConfigureJiraResponse"
        "500":
          $ref: "#/components/responses/InternalServerError"
        "501":
          $ref: "#/components/responses/NotImplemented"
  /projects/{projectId}/jira/oauth/start:
    post:
      tags: [Integrations]
      operationId: startProjectJiraOAuth
      summary: Start Atlassian OAuth for a project Jira connection.
      description: Returns the Atlassian authorization URL. In mock flow this route returns `501`.
      parameters:
        - $ref: "#/components/parameters/ProjectIdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JiraOAuthStartRequest"
            examples:
              startJiraOAuth:
                value:
                  personaId: priya
                  jiraSite: semicolons.atlassian.net
                  projectKey: SP
      responses:
        "200":
          description: Jira OAuth authorization URL.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JiraOAuthStartResponse"
        "500":
          $ref: "#/components/responses/InternalServerError"
        "501":
          $ref: "#/components/responses/NotImplemented"
  /jira/oauth/callback:
    get:
      tags: [Integrations]
      operationId: completeJiraOAuth
      summary: Complete Atlassian OAuth callback.
      parameters:
        - name: code
          in: query
          required: true
          schema:
            type: string
        - name: state
          in: query
          required: true
          schema:
            type: string
      responses:
        "302":
          description: Redirects back to the web app integrations page.
        "400":
          description: Missing OAuth code or state.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /projects/{projectId}/jira/sync:
    post:
      tags: [Integrations]
      operationId: syncProjectJira
      summary: Import Jira sprint issue signals for a project.
      description: Requires a completed Atlassian OAuth connection. In mock flow this route returns `501`.
      parameters:
        - $ref: "#/components/parameters/ProjectIdPath"
      requestBody:
        $ref: "#/components/requestBodies/PersonaActionBody"
      responses:
        "200":
          description: Jira sync completed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConfigureJiraResponse"
        "500":
          $ref: "#/components/responses/InternalServerError"
        "501":
          $ref: "#/components/responses/NotImplemented"
  /projects/{projectId}/git/configure:
    post:
      tags: [Integrations]
      operationId: configureProjectGit
      summary: Configure GitHub repository metadata for a project.
      description: In mock flow this route returns `501`.
      parameters:
        - $ref: "#/components/parameters/ProjectIdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConfigureGitRequest"
            examples:
              configureGit:
                value:
                  personaId: priya
                  provider: github
                  repoOwner: semicolons
                  repoName: sprintpulse
                  defaultBranch: main
      responses:
        "200":
          description: GitHub connection configured.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConfigureGitResponse"
        "500":
          $ref: "#/components/responses/InternalServerError"
        "501":
          $ref: "#/components/responses/NotImplemented"
  /projects/{projectId}/git/sync:
    post:
      tags: [Integrations]
      operationId: syncProjectGit
      summary: Import GitHub commit signals for a project.
      description: In mock flow this route returns `501`.
      parameters:
        - $ref: "#/components/parameters/ProjectIdPath"
      requestBody:
        $ref: "#/components/requestBodies/PersonaActionBody"
      responses:
        "200":
          description: Git sync completed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConfigureGitResponse"
        "500":
          $ref: "#/components/responses/InternalServerError"
        "501":
          $ref: "#/components/responses/NotImplemented"
  /projects/{projectId}/dashboard:
    get:
      tags: [Dashboard]
      operationId: getProjectDashboard
      summary: Get a project-scoped sprint health dashboard.
      parameters:
        - $ref: "#/components/parameters/ProjectIdPath"
        - $ref: "#/components/parameters/PersonaIdQuery"
        - $ref: "#/components/parameters/SprintIdQuery"
      responses:
        "200":
          description: Project dashboard.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectDashboardResponse"
        "404":
          description: Project dashboard is not found or not visible.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                notVisible:
                  value:
                    error: Project dashboard not found or not visible to this user
        "500":
          $ref: "#/components/responses/InternalServerError"
  /projects/{projectId}/ai/refresh:
    post:
      tags: [AI]
      operationId: refreshProjectAiInsights
      summary: Refresh AI-generated scoring and recommendations for a project.
      parameters:
        - $ref: "#/components/parameters/ProjectIdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [personaId]
              properties:
                personaId:
                  type: string
                sprintId:
                  type: string
      responses:
        "200":
          description: AI-enhanced project dashboard.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectDashboardResponse"
        "404":
          $ref: "#/components/responses/ProjectNotVisible"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /projects/{projectId}/ai/chat:
    post:
      tags: [AI]
      operationId: chatWithProjectAi
      summary: Ask the SprintPulse AI assistant about a project.
      parameters:
        - $ref: "#/components/parameters/ProjectIdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AiChatRequest"
      responses:
        "200":
          description: AI assistant response.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiChatResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/ProjectNotVisible"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
components:
  parameters:
    PersonaIdQuery:
      name: personaId
      in: query
      required: true
      description: Persona/profile id used by SprintPulse to resolve viewer permissions.
      schema:
        type: string
      example: priya
    SprintIdQuery:
      name: sprintId
      in: query
      required: false
      description: Optional sprint id filter for Supabase-backed sprint data.
      schema:
        type: string
      example: sprint-semi-2026
    ProjectIdPath:
      name: projectId
      in: path
      required: true
      description: Project id or, in mock flow, case-insensitive project key.
      schema:
        type: string
      example: sp-core
...
  requestBodies:
...
    PersonaActionBody:
      required: true
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/PersonaActionRequest"
          examples:
            actor:
              value:
                personaId: priya
...
  responses:
    InternalServerError:
      description: The route failed while loading or mutating data.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            internal:
              value:
                error: Unable to load data
    BadRequest:
      description: The request body or query is invalid.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotImplemented:
      description: Route is not available in the current runtime mode.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            notImplemented:
              value:
                error: Supabase data mode is enabled, but this route still needs its database-backed adapter.
...
    ProjectNotVisible:
      description: Project is not found or not visible to the viewer.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            projectNotVisible:
              value:
                error: Project not found or not visible to this user
...
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
          example: Persona not found
...
    Permission:
      type: string
      enum:
        - project:view
        - project:create
        - project:connect
        - project:editTeam
        - standup:submit
        - standup:sync
        - dashboard:viewTeam
        - dashboard:viewOwn
        - member:viewTeam
        - member:viewOwn
...
    Persona:
      type: object
      required: [id, name, email, initials, hackathonRole, productPersona, title, accessScope, focus]
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
          format: email
        initials:
          type: string
        hackathonRole:
          $ref: "#/components/schemas/HackathonRole"
        productPersona:
          $ref: "#/components/schemas/ProductPersona"
        title:
          type: string
        accessScope:
          $ref: "#/components/schemas/AccessScope"
        focus:
          type: string
...
    SprintProject:
      type: object
      required: [id, key, name, source, sprint, members, ownerIds, scrumMasterIds, createdBy, createdAt, updatedAt]
      properties:
        id:
          type: string
        key:
          type: string
        name:
          type: string
        source:
          $ref: "#/components/schemas/ProjectSource"
        jiraSite:
          type: string
        sprint:
          $ref: "#/components/schemas/SprintInfo"
        members:
          type: array
          items:
            $ref: "#/components/schemas/ProjectMember"
        ownerIds:
          type: array
          items:
            type: string
        scrumMasterIds:
          type: array
          items:
            type: string
        createdBy:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        lastSyncAt:
          type: string
          format: date-time
...
    SyncRun:
      type: object
      required: [id, projectId, source, status, requestedBy, startedAt, stats]
      properties:
        id:
          type: string
        projectId:
          type: string
        source:
          $ref: "#/components/schemas/SyncSource"
        status:
          $ref: "#/components/schemas/SyncRunStatus"
        requestedBy:
          type: string
        startedAt:
          type: string
          format: date-time
        finishedAt:
          type: string
          format: date-time
        stats:
          type: object
          additionalProperties:
            oneOf:
              - type: string
              - type: number
              - type: boolean
              - type: string
                nullable: true
        errorMessage:
          type: string
    SprintSummary:
      allOf:
        - $ref: "#/components/schemas/SprintInfo"
        - type: object
          required: [issueCount, standupCount, commitCount, blockerCount, healthScore]
          properties:
            issueCount:
              type: integer
              minimum: 0
            standupCount:
              type: integer
              minimum: 0
            commitCount:
              type: integer
              minimum: 0
            blockerCount:
              type: integer
              minimum: 0
            healthScore:
              type: integer
              minimum: 0
              maximum: 100
    JiraConnection:
      type: object
      required: [id, projectId, siteUrl, projectKey, status]
      properties:
        id:
          type: string
        projectId:
          type: string
        siteUrl:
          type: string
        projectKey:
          type: string
        status:
          $ref: "#/components/schemas/IntegrationStatus"
        cloudId:
          type: string
        displayName:
          type: string
        accountId:
          type: string
        boardId:
          type: integer
        activeSprintId:
          type: string
        activeSprintName:
          type: string
        authType:
          type: string
          enum: [manual, oauth]
        lastSyncAt:
          type: string
          format: date-time
        lastError:
          type: string
    GitConnection:
      type: object
      required: [id, projectId, provider, repoOwner, repoName, defaultBranch, status]
      properties:
        id:
          type: string
        projectId:
          type: string
        provider:
          type: string
          enum: [github]
        repoOwner:
          type: string
        repoName:
          type: string
        defaultBranch:
          type: string
        status:
          $ref: "#/components/schemas/IntegrationStatus"
        lastSyncAt:
          type: string
          format: date-time
    JiraIssue:
      type: object
      required: [id, projectId, issueKey, summary, status, daysIdle]
      properties:
        id:
          type: string
        projectId:
          type: string
        sprintId:
          type: string
        jiraIssueId:
          type: string
        issueKey:
          type: string
        summary:
          type: string
        status:
          $ref: "#/components/schemas/IssueStatus"
        assigneeProfileId:
          type: string
        jiraAssigneeId:
          type: string
        issueType:
          type: string
        priority:
          type: string
        url:
          type: string
          format: uri
        parentKey:
          type: string
        storyPoints:
          type: number
        daysIdle:
          type: integer
          minimum: 0
        storyPoints:
          type: number
          description: Jira story points when available, used to prioritize stale scope risk.
        updatedAtSource:
          type: string
          format: date-time
    GitCommit:
      type: object
      required: [id, projectId, sha, authorEmail, message, committedAt, additions, deletions]
      properties:
        id:
          type: string
        projectId:
          type: string
        sprintId:
          type: string
        sha:
          type: string
        authorProfileId:
          type: string
        authorEmail:
          type: string
          format: email
        message:
          type: string
        committedAt:
          type: string
          format: date-time
        additions:
          type: integer
          minimum: 0
        deletions:
          type: integer
          minimum: 0
...
    SprintListResponse:
      type: object
      required: [viewer, project, sprints]
      properties:
        viewer:
          $ref: "#/components/schemas/Persona"
        project:
          $ref: "#/components/schemas/SprintProject"
        currentSprint:
          $ref: "#/components/schemas/SprintSummary"
        sprints:
          type: array
          items:
            $ref: "#/components/schemas/SprintSummary"
...
    IntegrationStatusResponse:
      type: object
      required: [viewer, project, permissions, jira, git, recentRuns, issuePreview, commitPreview]
      properties:
        viewer:
          $ref: "#/components/schemas/Persona"
        project:
          $ref: "#/components/schemas/SprintProject"
        permissions:
          type: array
          items:
            $ref: "#/components/schemas/Permission"
        jira:
          type: object
          nullable: true
          allOf:
            - $ref: "#/components/schemas/JiraConnection"
        git:
          type: object
          nullable: true
          allOf:
            - $ref: "#/components/schemas/GitConnection"
        recentRuns:
          type: array
          items:
            $ref: "#/components/schemas/SyncRun"
        issuePreview:
          type: array
          items:
            $ref: "#/components/schemas/JiraIssue"
        commitPreview:
          type: array
          items:
            $ref: "#/components/schemas/GitCommit"
    ConfigureJiraRequest:
      type: object
      required: [personaId, jiraSite, projectKey]
      properties:
        personaId:
          type: string
        jiraSite:
          type: string
        projectKey:
          type: string
    JiraOAuthStartRequest:
      allOf:
        - $ref: "#/components/schemas/ConfigureJiraRequest"
    JiraOAuthStartResponse:
      type: object
      required: [authorizationUrl, state, expiresAt, warnings]
      properties:
        authorizationUrl:
          type: string
          format: uri
        state:
          type: string
        expiresAt:
          type: string
          format: date-time
        warnings:
          type: array
          items:
            type: string
    ConfigureJiraResponse:
      type: object
      required: [connection, importedIssues, warnings]
      properties:
        connection:
          $ref: "#/components/schemas/JiraConnection"
        run:
          $ref: "#/components/schemas/SyncRun"
        importedIssues:
          type: integer
          minimum: 0
        warnings:
          type: array
          items:
            type: string
    ConfigureGitRequest:
      type: object
      required: [personaId, provider, repoOwner, repoName]
      properties:
        personaId:
          type: string
        provider:
          type: string
          enum: [github]
        repoOwner:
          type: string
        repoName:
          type: string
        defaultBranch:
          type: string
    ConfigureGitResponse:
      type: object
      required: [connection, importedCommits, warnings]
      properties:
        connection:
          $ref: "#/components/schemas/GitConnection"
        run:
          $ref: "#/components/schemas/SyncRun"
        importedCommits:
          type: integer
          minimum: 0
        warnings:
          type: array
          items:
            type: string
...
    AiGenerationMeta:
      type: object
      required: [enabled, source, generatedAt]
      properties:
        enabled:
          type: boolean
        source:
          type: string
          enum: [openai, fallback, disabled, cache]
        generatedAt:
          type: string
          format: date-time
        model:
          type: string
        promptId:
          type: string
        reason:
          type: string
        cachedUntil:
          type: string
          format: date-time
...
    AiChatRequest:
      type: object
      required: [personaId, message]
      properties:
        personaId:
          type: string
        message:
          type: string
        sprintId:
          type: string
    AiChatResponse:
      type: object
      required: [answer, suggestedActions, meta]
      properties:
        answer:
          type: string
        suggestedActions:
          type: array
          items:
            type: string
        meta:
          $ref: "#/components/schemas/AiGenerationMeta"
...
    DashboardResponse:
      type: object
      required: [viewer, scope, summary, viewerPulse, memberPulses, teamPreview, recommendations]
      properties:
        viewer:
          $ref: "#/components/schemas/Persona"
        scope:
          $ref: "#/components/schemas/AccessScope"
        summary:
          $ref: "#/components/schemas/DashboardSummary"
        viewerPulse:
          $ref: "#/components/schemas/MemberPulse"
        memberPulses:
          type: array
          items:
            $ref: "#/components/schemas/MemberPulse"
        teamPreview:
          type: array
          items:
            $ref: "#/components/schemas/TeamPreviewItem"
        recommendations:
          type: array
          items:
            type: string
        ai:
          $ref: "#/components/schemas/AiDashboardOverlay"
        aiMeta:
          $ref: "#/components/schemas/AiGenerationMeta"
    ProjectDashboardResponse:
      allOf:
        - $ref: "#/components/schemas/DashboardResponse"
        - type: object
          required: [project]
          properties:
            project:
              $ref: "#/components/schemas/SprintProject"
...
    PersonaActionRequest:
      type: object
      required: [personaId]
      properties:
        personaId:
          type: string
...